added TelemetricVehicle
