+++ src/main/kotlin/com/micartel/telemetry/application/command/RegisterMileageHandler.kt
@@ -2,9 +2,10 @@ package com.micartel.telemetry.application.command
 
 import com.micartel.telemetry.domain.model.*
 
-class RegisterMileageHandler(private val vehicleRepository: VehicleRepository){
+class RegisterMileageHandler(private val vehicleRepository: TelemetricVehicleRepository){
     fun handle(c: RegisterMileageCommand) {
-        var vehicle = vehicleRepository.getByTelemetryDevice(SerialNumber(c.serialNumber))
+        val vehicle = vehicleRepository.get(SerialNumber(c.serialNumber))
+
         vehicle.changeMileage(Mileage(c.mileage, MileageUnit.KM))
         vehicleRepository.save(vehicle)
     }
