+++ src/main/kotlin/com/micartel/telemetry/domain/model/TelemetricVehicleRepository.kt
@@ -1,0 +1,6 @@
+package com.micartel.telemetry.domain.model
+
+interface TelemetricVehicleRepository {
+    fun save(vehicle: TelemetricVehicle)
+    fun get(device: SerialNumber): TelemetricVehicle
+}
