+++ src/main/kotlin/com/micartel/telemetry/domain/model/VehicleRepository.kt
@@ -3,5 +3,4 @@ package com.micartel.telemetry.domain.model
 interface VehicleRepository {
     fun save(vehicle: Vehicle)
     fun find(chassis: ChassisNumber): Vehicle?
-    fun getByTelemetryDevice(device: SerialNumber): Vehicle
 }
