+++ src/test/kotlin/com/micartel/telemetry/domain/mother/VehicleMother.kt
@@ -2,41 +2,56 @@ package com.micartel.telemetry.domain.mother
 
 import com.micartel.telemetry.domain.builder.VehicleRegistrar
 import com.micartel.telemetry.domain.model.*
+import com.micartel.telemetry.domain.service.assignTelemetryDevice
 
 const val CATEGORY_ELECTRIC = "CCAE"
 const val CATEGORY_HYBRID = "CCAI"
 const val CATEGORY_FUEL = "CCAD"
+val device:SerialNumber = SerialNumber("G-34567")
+
 class VehicleMother {
     companion object {
-        fun withMileage(mileage: Mileage) : Vehicle {
-            val vehicle = VehicleRegistrar().register()
+        fun with(chassis:ChassisNumber, device:SerialNumber): TelemetricVehicle{
+            return VehicleRegistrar()
+                .withChassis(chassis)
+                .register()
+                .assignTelemetryDevice(device)
+        }
+        fun fuel() : TelemetricVehicle {
+            return this.withCategory(Category(CATEGORY_FUEL))
+        }
+        fun electric() : TelemetricVehicle {
+            return this.withCategory(Category(CATEGORY_ELECTRIC))
+        }
+        fun hybrid() : TelemetricVehicle {
+            return this.withCategory(Category(CATEGORY_HYBRID))
+        }
+        fun withMileage(mileage: Mileage) : TelemetricVehicle {
+            val vehicle = this.create()
             vehicle.changeMileage(mileage)
             return vehicle
         }
-        fun withBatteryLevel(level: BatteryLevel) : Vehicle {
+        fun withBatteryLevel(level: BatteryLevel) : TelemetricVehicle {
             val vehicle = this.electric()
             vehicle.changeBatteryLevel(level)
             return vehicle
         }
-        fun withFuel(fuel: FuelLevel) : Vehicle {
+        fun withFuel(fuel: FuelLevel) : TelemetricVehicle {
             val vehicle = this.hybrid()
             vehicle.refuel(fuel)
             return vehicle
         }
-        fun fuel() : Vehicle {
+
+        private fun create(): TelemetricVehicle{
             return VehicleRegistrar()
-                .withCategory(Category(CATEGORY_FUEL))
                 .register()
+                .assignTelemetryDevice(device)
         }
-        fun electric() : Vehicle {
+        private fun withCategory(cat:Category): TelemetricVehicle{
             return VehicleRegistrar()
-                .withCategory(Category(CATEGORY_ELECTRIC))
+                .withCategory(cat)
                 .register()
-        }
-        fun hybrid() : Vehicle {
-            return VehicleRegistrar()
-                            .withCategory(Category(CATEGORY_HYBRID))
-                            .register()
+                .assignTelemetryDevice(device)
         }
     }
 }
